+++ uni_modules/uv-pick-color/components/uv-pick-color/props.js
@@ -42,11 +42,6 @@ export default {
 			type: String,
 			default: '#3c9cff'
 		},
-		// 滚动条的位置
-		scrollTop: {
-			type: [String, Number],
-			default: 0
-		},
 		...uni.$uv?.props?.pickColor
 	}
 }