+++ uni_modules/uv-ui-tools/libs/config/config.js
@@ -1,5 +1,5 @@
-// 此版本发布于2023-08-13
-const version = '1.1.4'
+// 此版本发布于2023-08-14
+const version = '1.1.5'
 
 // 开发环境才提示，生产环境不会提示
 if (process.env.NODE_ENV === 'development') {